Update snippet.filelister.php
When listed file is a directory, add query string parameter that identifies the directory path. This makes it possible to integrate more easily with the FileUpload snippet (switching &path dynamically based on a combination of the base path [unseen by users on front end] and the supplied directory path pulled from the query string).